CI: Release workflow (#237)
Previous PR for the same topic was
https://github.com/G-Research/spark-extension/pull/200, but closed as it
required major changes and approach which is shown in this PR

### **What type of PR is this?**

This is an improvement PR with the goal of enabling spark-extension
release using GitHub workflows

### **What this PR does / why we need it:**

The goal of this task is to move the release process to GitHub workflow
from current scripts that are run manually.
The current release process is documented in
[RELEASE.md](https://github.com/G-Research/spark-extension/blob/master/RELEASE.md/.
Due to its complexity, we reached out to one of the maintainers Enrico
who
[approved](https://github.com/G-Research/spark-extension/pull/200#issuecomment-1834357718)
this idea of moving `release.sh` to a github workflow for it's
transparent and better usage.

Resolves: https://github.com/G-Research/oss-portfolio-maturity/issues/32

### Acceptance criteria / Notes for the reviewer:

- [ ] [Comment on the issues from the maintainer
](https://github.com/G-Research/oss-portfolio-maturity/issues/32#issuecomment-1889050001)
- [ ] Keep the current process as is (not introducing breaking changes),
so the maintainers should always be able to do it the old way - but this
PR should introduce a workflow that will copy that behavior

Most of the comments and testing were done forked
[repo](https://github.com/gr-oss-devops/spark-extension/pull/2)
Part of maturity matrix:
https://github.com/G-Research/oss-portfolio-maturity/issues/32