Change environment for publishing snapshot to `snapshot` (#245)
Goal of this PR is to address
[issue](https://github.com/G-Research/spark-extension/pull/237#issuecomment-2220658888)
after merging https://github.com/G-Research/spark-extension/pull/237
The change we want to introduce is to use `snapshot` environment instead
of `release` which we use for publishing release. Also, we might not
want to use the same environment for snapshot and release releases.

TODO:
- [x] create new protected environment `snapshot` and set required rules
for publishing snapshot releases
- [x] copy & cleanup secrets from `release` env to `snapshot` env

---------

Co-authored-by: Enrico Minack <github@enrico.minack.dev>
Co-authored-by: Ivan Pavlovic <ivan@gr-oss.io>